fixed the update so now it deletes from uploads/images when making changes

+++ app/__init__.py
@@ -77,6 +77,22 @@ def save_image(image, image_name):
             print(f"FILE PATH: {file_path} doesn't exist.") 
     else:
         return None
+# Deletes upload and image from respective directories    
+def delete_from_uploads_and_images(id):
+    song = Song.query.get(id).filename
+    transformed_song = song.replace(' ', '_').replace(',', '').replace("'", '')
+    
+    root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
+    file_path = root + '/' + app.config['UPLOAD_FOLDER'] + '/' + transformed_song
+    os.remove(file_path)
+
+    image = Song.query.get(id).image
+    transformed_image = image.replace(' ', '_').replace(',', '').replace("'", '')
+
+    image_path = root + '/' + app.config['IMAGE_FOLDER'] + '/' + transformed_image
+    os.remove(image_path)
+    print(f"{song} was succesfully deleted from uploads w/ image.")
+    return f"{song} was successfully deleted."
 
 #Upload Route
 @app.route('/api/songs/upload', methods=['POST', 'GET'])
@@ -123,6 +139,8 @@ def update_song_information(id):
         return render_template('update_song.html', form=form)
     
     if request.method in ['PUT', 'POST'] and form.validate_on_submit():
+        delete_from_uploads_and_images(id)
+        
         file = request.files['filename']   
         filename = file.filename         
         file_path = save_song(file, filename)
@@ -166,27 +184,18 @@ def audio_player():
     file_names = os.listdir(file_path)
     return render_template('audio_player.html', file_names=file_names)
 
-#Delete song from /uploads and DB w/ associated image
-@app.route('/api/songs/<int:id>/delete', methods=['DELETE'])
-def delete_from_uploads_and_db(id):
-    song_for_db = Song.query.get(id)
-    song = Song.query.get(id).filename
-    transformed_song = song.replace(' ', '_').replace(',', '').replace("'", '')
-    
-    root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
-    file_path = root + '/' + app.config['UPLOAD_FOLDER'] + '/' + transformed_song
-    os.remove(file_path)
+# Delete song from /uploads and DB w/ associated image
+@app.route('/api/songs/<int:id>/delete', methods=['GET','DELETE'])
+def delete_song_from_db(id):
+    if request.method == "GET":
+        song_for_db = Song.query.get(id)
+        delete_from_uploads_and_images(id)
+        db.session.delete(song_for_db)
+        db.session.commit()
+        print(f"{song_for_db} was succesfully deleted from uploads w/ image and DB")
+        return f"{song_for_db} was successfully deleted from uploads w/ image and DB."
 
-    image = Song.query.get(id).image
-    transformed_image = image.replace(' ', '_').replace(',', '').replace("'", '')
 
-    image_path = root + '/' + app.config['IMAGE_FOLDER'] + '/' + transformed_image
-    os.remove(image_path)
-
-    db.session.delete(song_for_db)
-    db.session.commit()
-    print(f"{song} was succesfully deleted from uploads and from database with image.")
-    return f"{song} was successfully deleted."
         
 #Looping through uploads 
 @app.route('/api/config')
